Inline Pi05 private Gemma modules
