+++ docs_new/cookbook/vla/OpenPI/Pi0.5.mdx
@@ -271,8 +271,7 @@ Pi0.5 is implemented as a native VLA/action-policy pipeline, not as an SRT LLM s
 | VLA split group | `python/sglang/multimodal_gen/runtime/distributed/vla.py` | `VLASplitGroup`, `get_vla_split_group`, `broadcast_prefix_context` |
 | Denoise CUDA graph | `python/sglang/multimodal_gen/runtime/utils/vla_denoise_graph.py` | `VLADenoiseShapeBucket`, `VLADenoiseGraphRunner.capture_or_run` |
 | Pi0.5 policy model | `python/sglang/multimodal_gen/runtime/models/vlas/pi05_policy.py` | `Pi05PolicyModel.encode_prefix`, `Pi05PolicyModel.sample_actions`, `Pi05PolicyModel.denoise_step` |
-| Pi0.5 core modules | `python/sglang/multimodal_gen/runtime/models/vlas/pi05_core.py` | `Pi05CoreModel`, `PaliGemmaWithExpertModel` |
-| PiGemma / SigLIP native attention | `python/sglang/multimodal_gen/runtime/models/vlas/pi05_pigemma.py` | `PiGemmaModel`, `PiGemmaAttention`, `Pi05SiglipAttention`, `PiGemmaRMSNorm` |
+| Pi0.5 core modules | `python/sglang/multimodal_gen/runtime/models/vlas/pi05_core.py` | `Pi05CoreModel`, `PaliGemmaWithExpertModel`, `PiGemmaModel`, `PiGemmaAttention`, `Pi05SiglipAttention`, `PiGemmaRMSNorm` |
 
 `Pi05SamplingParams` inherits from the VLA/action-specific `VLASamplingParams`, not the image/video visual `SamplingParams`. The action params surface keeps robot-policy fields such as action horizon, action dimension, prefix cache, CUDA graph, observation, state, images, and optional action noise. Visual-generation fields such as height, width, fps, frame return, and negative prompt are intentionally absent. The base visual `SamplingParams._adjust` also dispatches `DataType.ACTION` before running image/video adjustment, so action requests do not inherit negative-prompt, resolution, frame-count, or visual sequence-shard logic.
